pyproject.toml
[tool.poetry]
name = "hcraft"
version = "0.0.0"
description = "H for Hephaestus."
authors = ["Toshinaki <shaduking@hotmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
pydantic = "^1.10.5"
pydash = "^6.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
